klzii_chat v0.0.1 KlziiChat.Router.Helpers

Module with named helpers generated from KlziiChat.Router.

Functions

auth_path(conn_or_endpoint, action)

auth_path(conn_or_endpoint, action, params)

auth_url(conn_or_endpoint, action)

auth_url(conn_or_endpoint, action, params)

banners_path(conn_or_endpoint, action)

banners_path(conn_or_endpoint, action, params)

banners_url(conn_or_endpoint, action)

banners_url(conn_or_endpoint, action, params)

chat_path(conn_or_endpoint, action)

chat_path(conn_or_endpoint, action, params)

chat_url(conn_or_endpoint, action)

chat_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

pinboard_resource_path(conn_or_endpoint, action)

pinboard_resource_path(conn_or_endpoint, action, params)

pinboard_resource_url(conn_or_endpoint, action)

pinboard_resource_url(conn_or_endpoint, action, params)

ping_path(conn_or_endpoint, action)

ping_path(conn_or_endpoint, action, params)

ping_url(conn_or_endpoint, action)

ping_url(conn_or_endpoint, action, params)

resources_path(conn_or_endpoint, action)

resources_path(conn_or_endpoint, action, params)

resources_path(conn_or_endpoint, action, id, params)

resources_url(conn_or_endpoint, action)

resources_url(conn_or_endpoint, action, params)

resources_url(conn_or_endpoint, action, id, params)

session_resources_path(conn_or_endpoint, action)

session_resources_path(conn_or_endpoint, action, params)

session_resources_path(conn_or_endpoint, action, id, params)

session_resources_url(conn_or_endpoint, action)

session_resources_url(conn_or_endpoint, action, params)

session_resources_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

surveys_path(conn_or_endpoint, action, id)

surveys_path(conn_or_endpoint, action, id, params)

surveys_url(conn_or_endpoint, action, id)

surveys_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.